+++ packages/granularity/src/components/GrDivider/config.ts
@@ -1,5 +1,8 @@
 import { defineGranularComponent } from '@feugene/unocss-preset-granular/contract'
 
+import { grDividerSafelist } from './safelist'
+
 export const grDividerConfig = defineGranularComponent(import.meta.url, {
   name: 'GrDivider',
+  safelist: grDividerSafelist,
 })