+++ packages/granularity/src/components/GrDivider/index.ts
@@ -1,4 +1,11 @@
 export { default } from './GrDivider.vue'
 export { default as GrDivider } from './GrDivider.vue'
-export type { GrDividerAlign, GrDividerOrientation } from './GrDivider.vue'
+export type {
+  GrDividerAlign,
+  GrDividerOrientation,
+  GrDividerSpacing,
+  GrDividerVariant,
+} from './GrDivider.vue'
+export type { GrDividerConfigurableProps } from './defaults'
 export { grDividerConfig } from './config'
+export { grDividerSafelist } from './safelist'